requirements.txt
certifi==2021.5.30
chardet==4.0.0
dnspython==2.1.0
python-dotenv==0.19.0
requests==2.25.1
tldextract==3.1.0
urllib3==1.26.6
python-whois==0.7.3
pyTelegramBotAPI==3.8.2
